move posthog-js to packages/browser
