+++ pnpm-workspace.yaml
@@ -1,5 +1,3 @@
 packages:
-  # all packages in direct subdirs of packages/
-  - "packages/*"
-  - "packages/browser/react"
-  - "eslint-rules"
+    # all packages in direct subdirs of packages/
+    - 'packages/*'
